+++ src/water/DRemoteTask.java
@@ -158,7 +158,7 @@ public abstract class DRemoteTask extends RemoteTask implements Cloneable {
     Key[] keys = null;
     if( val.type() == Value.ARRAYLET ) {
       ValueArray ary = (ValueArray)val;
-      keys = new Key[ary.chunks()];
+      keys = new Key[(int)ary.chunks()];
       for( int i=0; i<keys.length; i++ )
         keys[i] = ary.chunk_get(i);
     } else {
